pyproject.toml
[tool.poetry]
name = "hri-project"
version = "0.1.0"
description = ""
authors = ["Niclas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
autobahn = {extras = ["serialization", "twisted"], version = "^24.4.2"}
pyopenssl = "^25.0.0"
speechrecognition = "^3.14.2"
alpha-mini-rug = "^0.7.0"
google-generativeai = "^0.8.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
